Moves two functions from 'header' to '_util'
Also renames `_is_uncommentable` to `_is_commentable`,
to prevent needless double (and triple) negations,
which makes code harder to understand.
